# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount?: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  categoryId?: number;
  imageUrl?: string;
  isActive?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  stock?: number;
  categoryId?: number;
  imageUrl?: string;
  isActive?: boolean;
}

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  quantity: number;
  product?: Product;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CartItemUpdate {
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  address?: string | null;
  status: OrderStatus;
  total: number;
  createdAt: string;
}

export type OrderWithItemsStatus = typeof OrderWithItemsStatus[keyof typeof OrderWithItemsStatus];


export const OrderWithItemsStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderWithItems {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  address?: string | null;
  status: OrderWithItemsStatus;
  total: number;
  createdAt: string;
  items: OrderItem[];
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  address?: string;
  sessionId: string;
}

export type OrderStatusInputStatus = typeof OrderStatusInputStatus[keyof typeof OrderStatusInputStatus];


export const OrderStatusInputStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}

export interface DashboardSummary {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  totalCategories: number;
  pendingOrders: number;
  recentOrders: Order[];
}

export type ListProductsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetCartParams = {
sessionId: string;
};

export type ClearCartParams = {
sessionId: string;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
page?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;