# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Commerce API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: dashboard
    description: Dashboard analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart items
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Cleared

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"

  /cart/items/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details with items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard analytics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        productCount: { type: integer }

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }

    Product:
      type: object
      required: [id, name, price, stock, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        stock: { type: integer }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }

    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        categoryId: { type: integer }
        imageUrl: { type: string }
        isActive: { type: boolean, default: true }

    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        categoryId: { type: integer }
        imageUrl: { type: string }
        isActive: { type: boolean }

    CartItem:
      type: object
      required: [id, sessionId, productId, quantity]
      properties:
        id: { type: integer }
        sessionId: { type: string }
        productId: { type: integer }
        quantity: { type: integer }
        product:
          $ref: "#/components/schemas/Product"

    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId: { type: string }
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }

    OrderItem:
      type: object
      required: [id, orderId, productId, productName, price, quantity]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        price: { type: number }
        quantity: { type: integer }

    Order:
      type: object
      required: [id, customerName, customerEmail, status, total, createdAt]
      properties:
        id: { type: integer }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        total: { type: number }
        createdAt: { type: string, format: date-time }

    OrderWithItems:
      type: object
      required: [id, customerName, customerEmail, status, total, createdAt, items]
      properties:
        id: { type: integer }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        total: { type: number }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderInput:
      type: object
      required: [customerName, customerEmail, sessionId]
      properties:
        customerName: { type: string, minLength: 1 }
        customerEmail: { type: string, format: email }
        customerPhone: { type: string }
        address: { type: string }
        sessionId: { type: string }

    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]

    DashboardSummary:
      type: object
      required: [totalProducts, totalOrders, totalRevenue, totalCategories, recentOrders, pendingOrders]
      properties:
        totalProducts: { type: integer }
        totalOrders: { type: integer }
        totalRevenue: { type: number }
        totalCategories: { type: integer }
        pendingOrders: { type: integer }
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"