# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  sessionId: string;
  productId: number;
  /** @minimum 1 */
  quantity: number;
}