# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}