# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
}