# lib/api-zod/src/generated/types/orderStatusInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusInputStatus } from './orderStatusInputStatus';

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}